Add changes to implement Facade Pattern
