+++ ATMConsole/ATM/ITransaction.cs
@@ -1,0 +1,11 @@
+using ATMConsole;
+
+namespace ATMConsole
+{
+    interface ITransaction
+    {
+        void InsertTransaction(BankAccount bankAccount, Transaction transaction);
+
+        void ViewTransaction(BankAccount bankAccount);
+    }
+}
