+++ ATMConsole/Authentication/ILogin.cs
@@ -1,0 +1,8 @@
+namespace ATMConsole
+{
+    public interface ILogin
+    {
+        BankAccount CheckCardNoPassword(List<BankAccount> _accountList);
+
+    }
+}
