+++ ATMConsole/BankAccount/IBalance.cs
@@ -1,0 +1,7 @@
+namespace ATMConsole
+{
+    public interface IBalance
+    {
+        void CheckBalance(BankAccount bankAccount);
+    }
+}
