+++ ATMConsole/BankAccount/IDeposit.cs
@@ -1,0 +1,7 @@
+namespace ATMConsole
+{
+    public interface IDeposit
+    {
+        void PlaceDeposit(BankAccount bankAccount);
+    }
+}
