+++ ATMConsole/BankAccount/IThirdPartyTransfer.cs
@@ -1,0 +1,8 @@
+
+namespace ATMConsole
+{
+    public interface IThirdPartyTransfer
+    {
+        void PerformThirdPartyTransfer(BankAccount bankAccount, VMThirdPartyTransfer vmThirdPartyTransfer);
+    }
+}
