+++ ATMConsole/BankAccount/IWithdrawal.cs
@@ -1,0 +1,7 @@
+namespace ATMConsole
+{
+    public interface IWithdrawal
+    {
+        void MakeWithdrawal(BankAccount bankAccount);
+    }
+}
