--- ATMConsole/IBalance.cs
@@ -1,3 +1,0 @@
-public interface IBalance{
-    void CheckBalance(BankAccount bankAccount);
-}
