--- ATMConsole/IDeposit.cs
@@ -1,3 +1,0 @@
-public interface IDeposit{
-    void PlaceDeposit(BankAccount bankAccount);
-}
