--- ATMConsole/ILogin.cs
@@ -1,5 +1,0 @@
-public interface ILogin
-{
-    void CheckCardNoPassword();
-
-}
