--- ATMConsole/IThirdPartyTransfer.cs
@@ -1,3 +1,0 @@
-public interface IThirdPartyTransfer {
-    void PerformThirdPartyTransfer(BankAccount bankAccount,VMThirdPartyTransfer vmThirdPartyTransfer);
-}
