--- ATMConsole/ITransaction.cs
@@ -1,7 +1,0 @@
-using ATMConsole;
-
-interface ITransaction{
-    void InsertTransaction(BankAccount bankAccount, Transaction transaction);
-
-    void ViewTransaction(BankAccount bankAccount);
-}
