--- ATMConsole/IWithdrawal.cs
@@ -1,3 +1,0 @@
-public interface IWithdrawal{
-    void MakeWithdrawal(BankAccount bankAccount);
-}
